+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.List;
-import java.util.Set;
 
 public interface FilmStorage {
     Film save(Film film);
@@ -13,8 +12,6 @@ public interface FilmStorage {
     List<Film> getFilmList();
 
     void deleteAllFilms();
-
-    Set<Integer> getAllId();
 
     Film addFilm(Film film);
 }
