+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.filmorate.storage;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
-import java.util.Set;
 
 public interface UserStorage {
     User save(User user);
@@ -13,8 +12,6 @@ public interface UserStorage {
     List<User> getUserList();
 
     void deleteAllUsers();
-
-    Set<Integer> getAllId();
 
     User addUser(User user);
 }
